requirements.txt
PyYAML==5.4.1
feedparser==6.0.2
bs4==0.0.1
requests==2.25.1
dict_digger==0.2.1
python-dateutil==2.8.1